feat : PowerUp UI 및 골드 UI 설계

+++ Assets/Scripts/PowerUp/PowerUpScriptableObject.cs
@@ -7,7 +7,7 @@ public class PowerUpScriptableObject : ScriptableObject
     public Sprite IconSprite;
     public string powerUpName; // Display name
     public string description; // Tooltip or UI description
-    public float[] costPerLevel; // Base cost per level
+    public long[] costPerLevel; // Base cost per level
     public int CurrentLevel;
     public bool isPercentage = true; // if true add "%" in the tool tip
     public float[] upgradeValues; // Upgrade values for each level
